+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - SC2317: Warn about unreachable commands
 - SC2318: Warn about backreferences in 'declare x=1 y=$x'
 - SC2319/SC2320: Warn when $? refers to echo/printf/[ ]/[[ ]]/test
+- SC2321: Suggest removing $((..)) in array[$((idx))]=val
 
 ### Fixed
 - SC2086: Now uses DFA to make more accurate predictions about values